+++ clang/docs/ReleaseNotes.rst
@@ -288,6 +288,8 @@ Non-comprehensive list of changes in this release
 * Since MSVC 19.33 added undocumented attribute ``[[msvc::constexpr]]``, this release adds the attribute as well.
 
 * Added ``#pragma clang fp reciprocal``.
+* Builtins ``__builtin_shufflevector()`` and ``__builtin_convertvector()`` may now be used within constant
+  expressions.
 
 * The version of Unicode used by Clang (primarily to parse identifiers) has been updated to 15.1.